Added in post resource capabilities
For the sake of closing up development,
I made a resource upload workaround.
The frontend will make a post request with 1 file
to /posts/resources. They will get back a link
They will send this link with the rest
of a single post to /posts as a 'files' param.
This then becomes part of the post